+++ src/main/java/com/yesman/epicskills/server/commands/PlayerSkillTreeCommand.java
@@ -51,11 +51,18 @@ public class PlayerSkillTreeCommand {
 							Commands
 								.argument("targets", EntityArgument.players())
 								.executes(command -> {
-									return proccessSkillTreeCommand(command, EntityArgument.getPlayers(command, "targets"), null, null, Action.RESET, true, false);
+									return proccessSkillTreeCommand(command, EntityArgument.getPlayers(command, "targets"), null, null, Action.RESET, true, false, false);
 								})
+                                .then(
+                                    Commands
+                                        .argument("return_points", BoolArgumentType.bool())
+                                        .executes(command -> {
+                                            return proccessSkillTreeCommand(command, EntityArgument.getPlayers(command, "targets"), null, null, Action.RESET, true, false, BoolArgumentType.getBool(command, "return_points"));
+                                        })
+                                )
 						)
 						.executes(command -> {
-							return proccessSkillTreeCommand(command, ImmutableList.of(command.getSource().getPlayerOrException()), null, null, Action.RESET, true, false);
+							return proccessSkillTreeCommand(command, ImmutableList.of(command.getSource().getPlayerOrException()), null, null, Action.RESET, true, false, false);
 						})
 				)
 				.then(
@@ -81,7 +88,8 @@ public class PlayerSkillTreeCommand {
                                                                 command.getArgument("skill", Holder.class),
 																Action.UNLOCK,
 																BoolArgumentType.getBool(command, "force"),
-																false
+																false,
+                                                                false
 															);
 														})
 												)
@@ -93,7 +101,8 @@ public class PlayerSkillTreeCommand {
                                                         command.getArgument("skill", Holder.class),
 														Action.UNLOCK,
 														false,
-														false
+														false,
+                                                        false
 													);
 												})
 										)
@@ -105,7 +114,8 @@ public class PlayerSkillTreeCommand {
 												null,
 												Action.UNLOCK_TREE,
 												false,
-												false
+												false,
+                                                false
 											);
 										})
 								)
@@ -137,7 +147,8 @@ public class PlayerSkillTreeCommand {
 																		command.getArgument("skill", Holder.class),
 																		Action.LOCK,
 																		BoolArgumentType.getBool(command, "force"),
-																		BoolArgumentType.getBool(command, "unequip")
+																		BoolArgumentType.getBool(command, "unequip"),
+                                                                        false
 																	);
 																})
 														)
@@ -149,7 +160,8 @@ public class PlayerSkillTreeCommand {
 																command.getArgument("skill", Holder.class),
 																Action.LOCK,
 																BoolArgumentType.getBool(command, "force"),
-																false
+																false,
+                                                                false
 															);
 														})
 												)
@@ -161,7 +173,8 @@ public class PlayerSkillTreeCommand {
 														command.getArgument("skill", Holder.class),
 														Action.LOCK,
 														false,
-														false
+														false,
+                                                        false
 													);
 												})
 										)
@@ -179,7 +192,8 @@ public class PlayerSkillTreeCommand {
 																null,
 																Action.LOCK_TREE,
 																BoolArgumentType.getBool(command, "force"),
-																BoolArgumentType.getBool(command, "unequip")
+																BoolArgumentType.getBool(command, "unequip"),
+                                                                false
 															);
 														})
 												)
@@ -191,7 +205,8 @@ public class PlayerSkillTreeCommand {
 														null,
 														Action.LOCK_TREE,
 														BoolArgumentType.getBool(command, "force"),
-														false
+														false,
+                                                        false
 													);
 												})
 										)
@@ -203,7 +218,8 @@ public class PlayerSkillTreeCommand {
 												null,
 												Action.LOCK_TREE,
 												false,
-												false
+												false,
+                                                false
 											);
 										})
 								)
@@ -212,7 +228,16 @@ public class PlayerSkillTreeCommand {
 		);
 	}
 	
-	private static int proccessSkillTreeCommand(CommandContext<CommandSourceStack> command, Collection<ServerPlayer> players, @Nullable Holder.Reference<SkillTree> skillTree, @Nullable Holder<Skill> skill, Action action, boolean force, boolean unequip) throws CommandSyntaxException {
+	private static int proccessSkillTreeCommand(
+        CommandContext<CommandSourceStack> command,
+        Collection<ServerPlayer> players,
+        @Nullable Holder.Reference<SkillTree> skillTree,
+        @Nullable Holder<Skill> skill,
+        Action action,
+        boolean force,
+        boolean unequip,
+        boolean returnPoints
+    ) throws CommandSyntaxException {
 		int done = 0;
 		
 		if (players.isEmpty()) {
@@ -222,7 +247,7 @@ public class PlayerSkillTreeCommand {
 		switch (action) {
 		case RESET -> {
 			for (ServerPlayer player : players) {
-				if (resetTree(player)) {
+				if (resetTree(player, returnPoints)) {
 					done++;
 				}
 			}
@@ -292,9 +317,9 @@ public class PlayerSkillTreeCommand {
 		return done;
 	}
 	
-	private static boolean resetTree(ServerPlayer player) {
+	private static boolean resetTree(ServerPlayer player, boolean returnPoints) {
 		player.getExistingData(EpicSkillsAttachmentTypes.SKILL_TREE_PROGRESSION).ifPresent(skillTreeProgression -> {
-			skillTreeProgression.reload(false);
+			skillTreeProgression.reload(false, returnPoints);
 			EpicFightNetworkManager.sendToPlayer(new ClientBoundReloadSkillTree(false), player);
 		});